[TASK-200] home 에서 초대되도록 설정
